remix_full_without_measurements/garment-category-predictor/requirements.txt
flask
yacs
boto3
pillow
requests
opencv-python
torch
torchvision
transformers
numpy
safetensors
gunicorn